Draft: try support multi tags glog
